Factor out Python 3.8 compatibility
